Fix status error (#29)
* fix collatorSet get method

* Optimize the useCollatorByAddress hook by adding the inset parameter and adjusting the query logic; dynamically control the display of the Stake button in StakeManagementModal based on the collator's status.

* remove koi chain

* support unstakeRINGFromInactiveCollator and unstakeDepositsFromInactiveCollator

* fix ts error

* fix abi

* fix types error

+++ src/components/collator/_hooks/collator.ts
@@ -83,7 +83,7 @@ export const useCreateCollator = ({
 
   const { writeContractAsync, ...rest } = useWriteContract();
 
-  const oldKey = genKey({ address: address as `0x${string}`, votes: votes ?? 0n });
+  const oldKey = genKey({ address: address as `0x${string}`, votes: (votes as bigint) ?? 0n });
 
   const createCollator = useCallback(
     async ({ commission }: CreateAndCollatorProps) => {
